+++ backend/app/models/org.py
@@ -2,7 +2,7 @@
 from datetime import datetime
 from typing import Optional
 
-from sqlalchemy import BigInteger, DateTime, SmallInteger, String, func
+from sqlalchemy import BigInteger, DateTime, Integer, SmallInteger, String, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.models.base import Base, TimestampMixin
@@ -11,7 +11,7 @@ from app.models.base import Base, TimestampMixin
 class Department(TimestampMixin, Base):
     __tablename__ = "departments"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     parent_id: Mapped[Optional[int]] = mapped_column(BigInteger, nullable=True, index=True)
     name: Mapped[str] = mapped_column(String(64))
     leader_id: Mapped[Optional[int]] = mapped_column(BigInteger, nullable=True)
@@ -21,7 +21,7 @@ class Department(TimestampMixin, Base):
 class User(TimestampMixin, Base):
     __tablename__ = "users"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     username: Mapped[str] = mapped_column(String(64), unique=True)
     password_hash: Mapped[str] = mapped_column(String(128))
     display_name: Mapped[str] = mapped_column(String(64), default="")
@@ -33,7 +33,7 @@ class User(TimestampMixin, Base):
 class Role(TimestampMixin, Base):
     __tablename__ = "roles"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     role_name: Mapped[str] = mapped_column(String(64))
     role_code: Mapped[str] = mapped_column(String(64), unique=True)
     description: Mapped[str] = mapped_column(String(255), default="")
@@ -42,7 +42,7 @@ class Role(TimestampMixin, Base):
 class UserRole(Base):
     __tablename__ = "user_roles"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     user_id: Mapped[int] = mapped_column(BigInteger, index=True)
     role_id: Mapped[int] = mapped_column(BigInteger, index=True)
     created_at: Mapped["datetime"] = mapped_column(DateTime, server_default=func.now())
@@ -51,7 +51,7 @@ class UserRole(Base):
 class RolePermission(Base):
     __tablename__ = "role_permissions"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     role_id: Mapped[int] = mapped_column(BigInteger, index=True)
     permission_code: Mapped[str] = mapped_column(String(64), index=True)
     permission_type: Mapped[str] = mapped_column(String(16), default="api")
