+++ backend/app/models/settlement.py
@@ -5,13 +5,13 @@ from typing import Optional
 from sqlalchemy import JSON, BigInteger, DateTime, Integer, String, Text, func
 from sqlalchemy.orm import Mapped, mapped_column
 
-from app.models.base import Base, TimestampMixin
+from app.models.base import Base, BigIntPK, TimestampMixin
 
 
 class Faq(TimestampMixin, Base):
     __tablename__ = "faqs"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     question: Mapped[str] = mapped_column(String(512), index=True)
     answer: Mapped[str] = mapped_column(Text)
     category: Mapped[str] = mapped_column(String(64), default="")
@@ -26,7 +26,7 @@ class Faq(TimestampMixin, Base):
 class KnowledgeGap(TimestampMixin, Base):
     __tablename__ = "knowledge_gaps"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     question_pattern: Mapped[str] = mapped_column(String(512))
     sample_questions: Mapped[Optional[list]] = mapped_column(JSON, nullable=True)
     ask_count: Mapped[int] = mapped_column(Integer, default=1)
